backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
redis
celery==5.3.6
requests
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
